cache fetch accesses
we modify the environment interface so that it
looks up a page, rather than a specific fetch.

then we add a cache for fetches that not only covers
the MMU lookup, but the physical page lookup as well.

this improves performance by about 44%:
https://tech.lgbt/@mildsunrise/111710146324009126

I think what I'm doing here (having the cache skip
even the PTE bit set operation) is legal according to
RISC-V, since we're allowed to perform MMU lookup
speculatively.