✨ core idling!
implement the WFI instruction. to do that, core.h needs
to allow an environment to supply an implementation for it,
and to supply an implementation for the time counter.

then in the emulator, we still implement time as an
instruction counter, except when the emulator gets an WFI
and goes to sleep, in which case the time counter advances
according to actual time.

this kind of frankenstein allows a simpler implementation
of WFI (at least for now) and allows retaining most of the
deterministic properties of "instruction counter as time
counter".

since the emulator now needs to be aware of the clock freq
we set in the DTS, move that variable to a -D passed to both.