core: fix LR/SC
I think past Alba didn't quite get that stores *from the same hart* do not invalidate the reservation.  I'm not sure how this hasn't caused trouble yet, given that it prevents forward progress on any minimally complex LR/SC sequence (i.e. one that needed to use scratch memory).

this fact, combined with the allowance to make the reservation set as big as desired, means we don't actually have to remember the reservation set itself... we just have to know if it's active. and we
don't need to tap into stores at all, which simplifies the code